+++ src/main/java/units/stage/addressStage/LoadStage.java
@@ -67,7 +67,7 @@ public class LoadStage extends AddressStage {
 		reservedLoad++;
 	}
 
-	public float produce() {
+	public double produce() {
 		System.out.println("METHOD NOT IMPLEMENTED YET: LOAD -> PRODUCE");
 		return -1;
 	}
